fix: solved some bugs and improved the program

+++ tasks/internal/algorithms/mean.go
@@ -1,19 +1,19 @@
 package algorithms
 
 // GetMean returns the mean of a given field from a slice of tasks
-func GetMean(tasks *[]*Task, getFieldValue func(*Task) int64) float64 {
-	var sum int64
+func GetMean(tasks *[]*Task, getFieldValue func(*Task) float64) float64 {
+	var sum float64
 	for _, task := range *tasks {
 		sum += getFieldValue(task)
 	}
-	return float64(sum) / float64(len(*tasks))
+	return sum / float64(len(*tasks))
 }
 
 // GetMeanProcessingTime returns the mean processing time of a slice of tasks
 func GetMeanProcessingTime(tasks *[]*Task) float64 {
 	return GetMean(
-		tasks, func(task *Task) int64 {
-			return *task.processingTime
+		tasks, func(task *Task) float64 {
+			return float64(*task.processingTime)
 		},
 	)
 }
@@ -21,8 +21,8 @@ func GetMeanProcessingTime(tasks *[]*Task) float64 {
 // GetMeanWaitingTime returns the mean waiting time of a slice of tasks
 func GetMeanWaitingTime(tasks *[]*Task) float64 {
 	return GetMean(
-		tasks, func(task *Task) int64 {
-			return *task.waitingTime
+		tasks, func(task *Task) float64 {
+			return float64(*task.waitingTime)
 		},
 	)
 }
@@ -30,7 +30,7 @@ func GetMeanWaitingTime(tasks *[]*Task) float64 {
 // GetMeanServiceIndex returns the mean service index of a slice of tasks
 func GetMeanServiceIndex(tasks *[]*Task) float64 {
 	return GetMean(
-		tasks, func(task *Task) int64 {
+		tasks, func(task *Task) float64 {
 			return *task.serviceIndex
 		},
 	)
